Fixed up base_path generation in symlinks
My base_path generation logic wasn't great. When I was doing roles, I was able
to call the root `main.yml` and it would find the sub-yaml files and generate
the correct source path for symlinks.

This logic was broken when I tried to use `$PWD`, as it was actually for the
root `main.yml`. Now, I was going to append the subfolders, but this wouldn't
have worked if I called `main.yml` from a different folder, or called the
sub-yaml files directly.

Instead I went with the trick of getting the git repo's base path, and then
using that point of reference to do the relative paths.

Note: Would still be nice to find the "ansible" way of doing this.