Pulled in MonoDevelop IDE, although I've gotten emacs working, so may delete
Whilst I was waiting for MonoDevelop (stable) to be pulled in, I managed to get
emacs omnisharp completions to work. Ended up requiring a `.csproj` file in the
folder, which was created from: `dotnet new <project_type>`.

May not use MonoDevelop, but will keep it around for now, just in case.