Use the new official systemd suspend-then-hibernate.service
The `suspend-to-hibernate.service` in the arch wiki is no longer needed, since
there is now official `systemd` support. Found out that it was broken since I
had pulled in the latest changes (laptop on and hot in my bag :( ).

TODO: figure out how to fix the suspend key on the dell. Pressing it (Fn + End)
breaks any further `suspend-then-hibernate` attempts.